[master] Fixed cast error in unit test
[ 57%] Building CXX object
geometry/unit_tests/CMakeFiles/geometry_tests.dir/src/epsilon_algorithm_test.cpp.obj
d:/GJ/GIT/scientific_computing/ssc/geometry/unit_tests/src/epsilon_algorithm_test.cpp:
In member function 'virtual void
epsilon_algorithm_test_example_Test::TestBody()':
d:/GJ/GIT/scientific_computing/ssc/geometry/unit_tests/src/epsilon_algorithm_test.cpp:36:38:
error: conversion to 'double' from 'size_t {aka long long unsigned int}'
may alter its value [-Werror=conversion]
         S.push_back(S.back()+pow(-1,i)/(2.*i+1));
	                                       ^
					       compilation terminated due
					       to -Wfatal-errors.
					       cc1plus.exe: all warnings
					       being treated as errors

+++ ssc/geometry/unit_tests/src/epsilon_algorithm_test.cpp
@@ -33,7 +33,7 @@ TEST_F(epsilon_algorithm_test, example)
     const size_t N = 25;
     for (size_t i = 1 ; i < N ; ++i)
     {
-        S.push_back(S.back()+pow(-1,i)/(2.*i+1));
+        S.push_back(S.back()+pow(-1,(double)i)/(2.*(double)i+1.));
     }
 //! [epsilon_algorithm_test example]
 //! [epsilon_algorithm_test expected output]
